version upgrades - spring* and java

+++ adapter/src/main/java/play/with/integration/batch/model/Count.java
@@ -1,18 +1,18 @@
 package play.with.integration.batch.model;
 
-import org.springframework.batch.core.StepExecution;
+import org.springframework.batch.core.step.StepExecution;
 import play.with.integration.batch.util.JsonUtils;
 
 import java.io.Serializable;
 
 public class Count implements Serializable {
-    private int readCount;
-    private int readSkipCount;
-    private int processSkipCount;
-    private int writeCount;
-    private int writeSkipCount;
+    private long readCount;
+    private long readSkipCount;
+    private long processSkipCount;
+    private long writeCount;
+    private long writeSkipCount;
 
-    public Count(int readCount, int readSkipCount, int processSkipCount, int writeCount, int writeSkipCount) {
+    public Count(long readCount, long readSkipCount, long processSkipCount, long writeCount, long writeSkipCount) {
         this.readCount = readCount;
         this.readSkipCount = readSkipCount;
         this.processSkipCount = processSkipCount;
@@ -20,43 +20,43 @@ public class Count implements Serializable {
         this.writeSkipCount = writeSkipCount;
     }
 
-    public int getReadCount() {
+    public long getReadCount() {
         return readCount;
     }
 
-    public void setReadCount(int readCount) {
+    public void setReadCount(long readCount) {
         this.readCount = readCount;
     }
 
-    public int getReadSkipCount() {
+    public long getReadSkipCount() {
         return readSkipCount;
     }
 
-    public void setReadSkipCount(int readSkipCount) {
+    public void setReadSkipCount(long readSkipCount) {
         this.readSkipCount = readSkipCount;
     }
 
-    public int getProcessSkipCount() {
+    public long getProcessSkipCount() {
         return processSkipCount;
     }
 
-    public void setProcessSkipCount(int processSkipCount) {
+    public void setProcessSkipCount(long processSkipCount) {
         this.processSkipCount = processSkipCount;
     }
 
-    public int getWriteCount() {
+    public long getWriteCount() {
         return writeCount;
     }
 
-    public void setWriteCount(int writeCount) {
+    public void setWriteCount(long writeCount) {
         this.writeCount = writeCount;
     }
 
-    public int getWriteSkipCount() {
+    public long getWriteSkipCount() {
         return writeSkipCount;
     }
 
-    public void setWriteSkipCount(int writeSkipCount) {
+    public void setWriteSkipCount(long writeSkipCount) {
         this.writeSkipCount = writeSkipCount;
     }
